fix: :bug: change Environment.TickCount64 to Environment.TickCount

+++ logic/Preparation/Interface/IDoorway.cs
@@ -4,7 +4,7 @@ namespace Preparation.Interface
 {
     public interface IDoorway : IGameObj
     {
-        public int OpenStartTime { get; }
+        public long OpenStartTime { get; }
         public int OpenDegree { get; }
         public bool StopOpenning();
         public bool TryToOpen();
